code/ui/requirements.txt
pygame
pyserial